Allow a fresh clone for some test cases

+++ src/main/java/reposense/report/RepoCloner.java
@@ -51,6 +51,8 @@ public class RepoCloner {
     private static final int MAX_NO_OF_REPOS = 2;
     private static final Logger logger = LogsManager.getLogger(RepoCloner.class);
 
+    private static final boolean DEFAULT_IS_CLONING_FORCED = false;
+
     private RepoConfiguration[] configs = new RepoConfiguration[MAX_NO_OF_REPOS];
     private int currentIndex = 0;
     private int previousIndex = 0;
@@ -63,25 +65,34 @@ public class RepoCloner {
      * Does not wait for process to finish executing.
      */
     public void cloneBare(RepoConfiguration config) {
+        cloneBare(config, DEFAULT_IS_CLONING_FORCED);
+    }
+
+
+    /**
+     * Spawns a process to clone the bare repository specified by {@code config}.
+     * Does not wait for process to finish executing.
+     */
+    public void cloneBare(RepoConfiguration config, boolean isCloningForTestForced) {
         configs[currentIndex] = config;
 
         if (!config.isShallowCloningPerformed()) {
-            isCurrentRepoCloned = spawnCloneProcess(config);
+            isCurrentRepoCloned = spawnCloneProcess(config, isCloningForTestForced);
         } else {
-            boolean didShallowPartialCloneSucceed = spawnShallowPartialCloneProcess(config);
+            boolean didShallowPartialCloneSucceed = spawnShallowPartialCloneProcess(config, isCloningForTestForced);
             String shallowPartialBareRoot = FileUtil.getShallowPartialBareRepoPath(config).toString();
 
             if (!didShallowPartialCloneSucceed || GitRevList.checkIsEmptyRepo(shallowPartialBareRoot)) {
-                isCurrentRepoCloned = spawnCloneProcess(config);
+                isCurrentRepoCloned = spawnCloneProcess(config, isCloningForTestForced);
                 return;
             }
 
             List<String> graftedCommits = GitRevList.getRootCommits(shallowPartialBareRoot);
             List<String> graftedCommitParents = GitCatFile.getParentsOfCommits(shallowPartialBareRoot, graftedCommits);
 
-            boolean didPartialCloneSucceed = spawnPartialCloneProcess(config);
+            boolean didPartialCloneSucceed = spawnPartialCloneProcess(config, isCloningForTestForced);
             if (!didPartialCloneSucceed) {
-                isCurrentRepoCloned = spawnCloneProcess(config);
+                isCurrentRepoCloned = spawnCloneProcess(config, isCloningForTestForced);
                 return;
             }
 
@@ -94,9 +105,9 @@ public class RepoCloner {
                 sinceDate = null;
             }
             if (sinceDate != null) {
-                isCurrentRepoCloned = spawnShallowCloneProcess(config, sinceDate);
+                isCurrentRepoCloned = spawnShallowCloneProcess(config, sinceDate, isCloningForTestForced);
             } else {
-                isCurrentRepoCloned = spawnCloneProcess(config);
+                isCurrentRepoCloned = spawnCloneProcess(config, isCloningForTestForced);
             }
         }
     }
@@ -149,7 +160,7 @@ public class RepoCloner {
      * Spawns a process to clone repo specified in {@code config}. Does not wait for process to finish executing.
      * Should only handle a maximum of one spawned process at any time.
      */
-    private boolean spawnCloneProcess(RepoConfiguration config) {
+    private boolean spawnCloneProcess(RepoConfiguration config, boolean isCloningForTestForced) {
         assert(crp == null);
 
         try {
@@ -160,8 +171,9 @@ public class RepoCloner {
             Path rootPath = Paths.get(FileUtil.REPOS_ADDRESS, config.getRepoFolderName());
             Path repoPath = Paths.get(rootPath.toString(), config.getRepoName());
 
-            if (SystemUtil.isTestEnvironment() && Files.exists(repoPath)) {
-                logger.info("Skipped cloning from " + config.getLocation() + " as it was cloned before.");
+            if (SystemUtil.isTestEnvironment() && Files.exists(repoPath) && !isCloningForTestForced) {
+                logger.info("Skipped cloning from " + config.getLocation()
+                        + " as it was cloned before and cloning is not forced.");
                 return true;
             }
 
@@ -181,12 +193,24 @@ public class RepoCloner {
      * Does not wait for process to finish executing.
      * Should only handle a maximum of one spawned process at any time.
      */
-    private boolean spawnShallowCloneProcess(RepoConfiguration config, Date shallowSinceDate) {
+    private boolean spawnShallowCloneProcess(RepoConfiguration config, Date shallowSinceDate,
+                                             boolean isCloningForTestForced) {
         assert(crp == null);
 
         try {
-            FileUtil.deleteDirectory(FileUtil.getBareRepoPath(config).toString());
+            if (!SystemUtil.isTestEnvironment()) {
+                FileUtil.deleteDirectory(FileUtil.getBareRepoPath(config).toString());
+            }
+
             Path rootPath = Paths.get(FileUtil.REPOS_ADDRESS, config.getRepoFolderName());
+            Path repoPath = Paths.get(rootPath.toString(), config.getRepoName());
+
+            if (SystemUtil.isTestEnvironment() && Files.exists(repoPath) && !isCloningForTestForced) {
+                logger.info("Skipped cloning from " + config.getLocation()
+                        + " as it was cloned before and cloning is not forced.");
+                return true;
+            }
+
             Files.createDirectories(rootPath);
 
             logger.info(String.format(MESSAGE_START_CLONING_SHALLOW, config.getLocation()));
@@ -203,12 +227,23 @@ public class RepoCloner {
      * Spawns a process to create partial clone of repo specified in {@code config}.
      * Waits for process to finish executing.
      */
-    private boolean spawnPartialCloneProcess(RepoConfiguration config) {
+    private boolean spawnPartialCloneProcess(RepoConfiguration config, boolean isCloningForTestForced) {
         assert(crp == null);
 
         try {
-            FileUtil.deleteDirectory(FileUtil.getPartialBareRepoPath(config).toString());
+            if (!SystemUtil.isTestEnvironment()) {
+                FileUtil.deleteDirectory(FileUtil.getPartialBareRepoPath(config).toString());
+            }
+
             Path rootPath = Paths.get(FileUtil.REPOS_ADDRESS, config.getRepoFolderName());
+            Path repoPath = Paths.get(rootPath.toString(), config.getRepoName());
+
+            if (SystemUtil.isTestEnvironment() && Files.exists(repoPath) && !isCloningForTestForced) {
+                logger.info("Skipped cloning from " + config.getLocation()
+                        + " as it was cloned before and cloning is not forced.");
+                return true;
+            }
+
             Files.createDirectories(rootPath);
 
             logger.info(String.format(MESSAGE_START_CLONING_PARTIAL, config.getLocation()));
@@ -224,12 +259,22 @@ public class RepoCloner {
      * Spawns a process to create shallow partial clone of repo specified in {@code config}.
      * Waits for process to finish executing.
      */
-    private boolean spawnShallowPartialCloneProcess(RepoConfiguration config) {
+    private boolean spawnShallowPartialCloneProcess(RepoConfiguration config, boolean isCloningForTestForced) {
         assert(crp == null);
 
         try {
-            FileUtil.deleteDirectory(FileUtil.getShallowPartialBareRepoPath(config).toString());
+            if (!SystemUtil.isTestEnvironment()) {
+                FileUtil.deleteDirectory(FileUtil.getShallowPartialBareRepoPath(config).toString());
+            }
             Path rootPath = Paths.get(FileUtil.REPOS_ADDRESS, config.getRepoFolderName());
+            Path repoPath = Paths.get(rootPath.toString(), config.getRepoName());
+
+            if (SystemUtil.isTestEnvironment() && Files.exists(repoPath) && !isCloningForTestForced) {
+                logger.info("Skipped cloning from " + config.getLocation()
+                        + " as it was cloned before and cloning is not forced.");
+                return true;
+            }
+
             Files.createDirectories(rootPath);
 
             logger.info(String.format(MESSAGE_START_CLONING_SHALLOW_PARTIAL, config.getLocation()));
